+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Add style support for radio and checkbox in Firefox
+- Add `autoplay` setting to Lightbox
+- Add `poster` setting to Lightbox items
 
 ### Changed
 
@@ -12,13 +14,16 @@
 - Remove `uk-section-media` from Section component
 - Change `uk-tab-left` text alignment to left
 - Update overview in tests
-- Rename 'isWithin' to 'within'
+- Rename `isWithin` to `within`
+- Allow fullscreen mode for YouTube and Vimeo videos in Lightbox
+- Video component now stops playing the video regardsless of the autoplay setting
 
 ### Fixed
 
 - Fix responsive images in modal for IE11
 - Fix close button outside causing a scrollbar on small devices
 - Fix inverse style for `uk-hr`
+- Fix video size of Vimeo videos in Lightbox component
 
 ## 3.0.0 beta 30 (August 18, 2017)